birthday card refactor
