--- lib/components/birthday_card.dart
@@ -1,245 +1,0 @@
-import 'dart:async';
-
-import 'package:daty/screens/birthday_info_page.dart';
-import 'package:daty/utilities/Birthday.dart';
-import 'package:daty/utilities/calculator.dart';
-import 'package:daty/utilities/constants.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-
-class BirthdayCard extends StatefulWidget {
-  final Birthday birthday;
-  final bool canTap;
-
-  const BirthdayCard(this.birthday, this.canTap, {super.key});
-
-  @override
-  State<BirthdayCard> createState() => _BirthdayCardState();
-}
-
-class _BirthdayCardState extends State<BirthdayCard>
-    with WidgetsBindingObserver {
-  late Timer timer;
-  late Duration duration;
-
-  void startTimer() {
-    duration = Duration(
-      hours: Calculator.hoursTillBirthday(widget.birthday.date),
-      minutes: Calculator.minutesTillBirthday(widget.birthday.date),
-      seconds: Calculator.secondsTillBirthday(widget.birthday.date),
-    );
-
-    timer = Timer.periodic(
-      duration,
-      (Timer t) => mounted
-          ? setState(() {
-              duration = const Duration(hours: 24);
-            })
-          : timer.cancel(),
-    );
-  }
-
-  @override
-  void initState() {
-    super.initState();
-
-    WidgetsBinding.instance.addObserver(this);
-    startTimer();
-  }
-
-  @override
-  void dispose() {
-    timer.cancel();
-
-    WidgetsBinding.instance.removeObserver(this);
-    super.dispose();
-  }
-
-  @override
-  void didChangeAppLifecycleState(AppLifecycleState state) {
-    switch (state) {
-      case AppLifecycleState.paused:
-        timer.cancel();
-        break;
-      case AppLifecycleState.inactive:
-        timer.cancel();
-        break;
-      case AppLifecycleState.resumed:
-        timer = Timer.periodic(duration, (Timer t) => setState(() {}));
-        break;
-      default:
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () async {
-        if (widget.canTap) {
-          await Navigator.of(context).push(
-            MaterialPageRoute(
-              builder: (BuildContext context) {
-                return BirthdayInfoPage(widget.birthday.birthdayId);
-              },
-            ),
-          );
-        }
-      },
-      child: Container(
-        padding: const EdgeInsets.all(15.0),
-        decoration: const BoxDecoration(
-          color: Constants.greySecondary,
-          borderRadius: BorderRadius.all(Radius.circular(20)),
-        ),
-        child: Row(
-          mainAxisSize: MainAxisSize.max,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            birthdayDisplay(),
-            const SizedBox(width: 15),
-            birthdayInfo(),
-            const Spacer(),
-            Calculator.hasBirthdayToday(widget.birthday.date)
-                ? partyIcon()
-                : dayCounter(),
-            Container(
-              padding: const EdgeInsets.only(left: 10.0),
-              child: widget.canTap
-                  ? const Icon(
-                      Icons.arrow_forward_ios_rounded,
-                      color: Constants.whiteSecondary,
-                      size: 20,
-                    )
-                  : null,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Container dayCounter() {
-    return Container(
-      decoration: const BoxDecoration(
-        color: Constants.bluePrimary,
-        shape: BoxShape.circle,
-        boxShadow: [
-          BoxShadow(
-            color: Constants.bluePrimary,
-            spreadRadius: 1,
-            blurRadius: 8,
-          ),
-        ],
-      ),
-      width: 60,
-      height: 60,
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          Text(
-            Calculator.remainingDaysTillBirthday(widget.birthday.date)
-                .toString(),
-            style: const TextStyle(
-                fontSize: Constants.normalFontSize,
-                fontWeight: FontWeight.bold,
-                color: Constants.whiteSecondary),
-          ),
-          Text(
-            AppLocalizations.of(context)!.days,
-            style: const TextStyle(
-              fontSize: Constants.smallerFontSize,
-              fontWeight: FontWeight.bold,
-              color: Constants.whiteSecondary,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Container partyIcon() {
-    return Container(
-      decoration: const BoxDecoration(
-        color: Constants.purplePrimary,
-        shape: BoxShape.circle,
-        boxShadow: [
-          BoxShadow(
-            color: Constants.purplePrimary,
-            spreadRadius: 1,
-            blurRadius: 8,
-          ),
-        ],
-      ),
-      width: 60,
-      height: 60,
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: const [
-          Text(
-            '🎉',
-            style: TextStyle(
-              fontSize: Constants.titleFontSizeSize,
-              fontWeight: FontWeight.bold,
-              shadows: [
-                Shadow(
-                  color: Constants.blackPrimary,
-                  offset: Offset(0.3, 0.2),
-                  blurRadius: 15.0,
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Column birthdayInfo() {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          widget.birthday.name,
-          style: const TextStyle(
-            fontSize: Constants.biggerFontSize,
-            fontWeight: FontWeight.bold,
-            color: Constants.whiteSecondary,
-          ),
-        ),
-        Text(
-          '${Calculator.getDayName(widget.birthday.date.weekday, context)}, ${widget.birthday.date.day}. ${Calculator.getMonthName(widget.birthday.date.month, context)}',
-          style: const TextStyle(
-            fontSize: Constants.smallerFontSize,
-            color: Constants.whiteSecondary,
-          ),
-        ),
-      ],
-    );
-  }
-
-  Column birthdayDisplay() {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        const Icon(
-          Icons.cake_outlined,
-          size: 30,
-          color: Constants.whiteSecondary,
-        ),
-        Text(
-          Calculator.hasBirthdayToday(widget.birthday.date)
-              ? '${Calculator.calculateAge(widget.birthday.date)}'
-              : '${Calculator.calculateAge(widget.birthday.date) + 1}',
-          style: const TextStyle(
-            fontSize: Constants.biggerFontSize,
-            fontWeight: FontWeight.bold,
-            color: Constants.whiteSecondary,
-          ),
-        ),
-      ],
-    );
-  }
-}
